add comments and refactor
